refactor: Move type definitions to seperate files
