+++ src/components/Layout/Layout.tsx
@@ -1,11 +1,8 @@
-import { FC, ReactNode } from 'react';
-
-interface LayoutInterface {
-	children: ReactNode;
-}
+import { FC } from 'react';
+import { LayoutInterface } from '../../types/layout';
 
 export const Layout: FC<LayoutInterface> = ({ children }) => (
 	<div className="layout">
 		{children}
 	</div>
-)
+);
