+++ src/types/events.ts
@@ -1,5 +1,3 @@
 import { MouseEvent } from 'react';
 
-// fix event unused vars tslint warning
-// eslint-disable-next-line 
 export type MouseEventHandler = (event: MouseEvent<HTMLElement>) => void;
