+++ src/types/list.ts
@@ -1,0 +1,17 @@
+import { ReactNode } from 'react';
+
+export interface ListInterface {
+	children?: ReactNode;
+	extraClass?: string;
+}
+
+export interface ListItemInterface {
+	children?: ReactNode;
+	extraClass?: string;
+	jobResponsibility?: string;
+}
+
+export interface ListContentInterface {
+	children?: ReactNode;
+	text?: string;
+}
