+++ src/types/progress.ts
@@ -1,0 +1,3 @@
+export interface ProgressBarInterface {
+	progress?: number;
+}
